refactor: PlaceDetailCollectionHandler 생성하여 PlaceDetailViewController가 비대해지는 문제 해결
