+++ iOS_Wegg/iOS_Wegg/App/SceneDelegate.swift
@@ -26,7 +26,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         navigationController.setNavigationBarHidden(true, animated: false)
         
         // 원하는 뷰 컨트롤러 파일의 이름을 작성
-        window?.rootViewController = PlaceDetailViewController()
+        window?.rootViewController = MainTabBarController()
         window?.makeKeyAndVisible()
     }
