+++ iOS_Wegg/iOS_Wegg/Screen/Map/PlaceDetail/ViewController/PlaceDetailViewController.swift
@@ -8,7 +8,9 @@
 import UIKit
 
 class PlaceDetailViewController: UIViewController {
-    var targetSectionIndex: Int = 0 // 모델에서 원하는 인덱스 설정
+    var targetSectionIndex: Int = 3 // 모델에서 원하는 인덱스 설정
+    private var collectionHandler: PlaceDetailCollectionHandler?
+    private let sampleSections = HotPlaceSectionModel.sampleSections
     private var detailData: HotPlaceDetailModel?
     private lazy var placeDetailView = PlaceDetailView()
 
@@ -21,8 +23,12 @@ class PlaceDetailViewController: UIViewController {
     }
     
     private func configureCollectionView() {
-        placeDetailView.studyImageCollectionView.delegate = self
-        placeDetailView.studyImageCollectionView.dataSource = self
+        collectionHandler = PlaceDetailCollectionHandler(
+            targetSectionIndex: targetSectionIndex,
+            sampleSections: sampleSections
+        )
+        placeDetailView.studyImageCollectionView.delegate = collectionHandler
+        placeDetailView.studyImageCollectionView.dataSource = collectionHandler
         placeDetailView.gestureDelegate = self
     }
     
@@ -71,77 +77,6 @@ class PlaceDetailViewController: UIViewController {
         ]
         
         info.forEach { $0.key.text = $0.value }
-    }
-}
-
-// MARK: - UICollectionViewDelegate & UICollectionViewDataSource
-
-extension PlaceDetailViewController:
-    UICollectionViewDelegateFlowLayout,
-    UICollectionViewDataSource {
-    
-    /// 컬렉션 뷰에서 하나의 섹션만 표시
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    /// 특정 인덱스(`targetSectionIndex`)에 해당하는 섹션의 아이템 개수를 반환
-    func collectionView(
-        _ collectionView: UICollectionView,
-        numberOfItemsInSection section: Int
-    ) -> Int {
-        return HotPlaceSectionModel.sampleSections[
-            safe: targetSectionIndex
-        ]?.items.count ?? 0
-    }
-    
-    /// 특정 인덱스(`targetSectionIndex`)의 섹션에서 셀을 생성하고 데이터를 설정
-    func collectionView(
-        _ collectionView: UICollectionView,
-        cellForItemAt indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(
-            withReuseIdentifier: PlaceDetailImageCell.identifier,
-            for: indexPath
-        ) as? PlaceDetailImageCell else {
-            fatalError("Could not dequeue PlaceDetailImageCell")
-        }
-        
-        // safe를 통해`targetSectionIndex`와 `indexPath.row`가 안전한 범위 내에 있는 경우만 실행
-        if let section = HotPlaceSectionModel.sampleSections[safe: targetSectionIndex],
-           let data = section.items[safe: indexPath.row] {
-            cell.configure(model: data)
-        }
-        
-        return cell
-    }
-    
-    // 컬렉션 뷰에서 행(줄) 간 간격 설정
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        minimumLineSpacingForSectionAt section: Int
-    ) -> CGFloat {
-        return 14
-    }
-    
-    // 같은 행(라인) 내에서 아이템 간 최소 간격 설정
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        minimumInteritemSpacingForSectionAt section: Int
-    ) -> CGFloat {
-        return 0
-    }
-    
-    // 컬렉션 뷰의 특정 아이템이 선택되었을 때 실행되는 이벤트 처리
-    func collectionView(
-        _ collectionView: UICollectionView,
-        didSelectItemAt indexPath: IndexPath
-    ) {
-        if indexPath.section == 0 && indexPath.item == 0 {
-            print("이미지 컬렉션뷰 탭")
-        }
     }
 }
